requirements.txt
Flask
flask-cors
gunicorn
sentence-transformers
faiss-cpu==1.6.5
prometheus-client
prometheus-flask-exporter
